requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.1
python-dotenv==1.0.1
Pillow==10.1.0
boto3==1.34.69
pymongo==4.8.0
pymysql==1.1.0
cryptography==41.0.7
redis~=5.0
PyJWT==2.8.0
